1) заменил If'ы на case
